Refactor dnit/infrastructure_utils into smaller files and better interfaces
