+++ dnit/deps.ts
@@ -1,0 +1,25 @@
+// task execution tool - dnit:
+export {
+  task,
+  exec as dnitExec,
+  file as trackFile,
+  runAlways,
+  path,
+  fs,
+  Task,
+  TrackedFile,
+} from 'https://deno.land/x/dnit@dnit-v1.8.0/mod.ts';
+
+// hx dnit-utils:
+export {
+  runConsole,
+  dockerRunConsole,
+  dockerRun,
+  currentUserOpts,
+  confirmation,
+} from 'https://denopkg.com/helix-collective/dnit-utils@v1.2.0/mod.ts';
+
+// other deno typescript libs:
+export * as jszip from 'https://deno.land/x/jszip@0.7.0/mod.ts';
+
+export * as changeCase from 'https://deno.land/x/case@v2.1.0/mod.ts';
